# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MyBank API",
    "description": "API per progetto universitario bancario MyBank sviluppato con Next.js",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/"
    }
  ],
...
  "paths": {
...
    "/transactions": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Recupera storico transazioni dell'utente con filtri opzionali",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "schema": {
              "type": "integer"
            },
            "description": "ID conto da filtrare (opzionale)"
          },
          {
            "in": "query",
            "name": "transaction_type",
            "schema": {
              "type": "string",
              "enum": ["deposit", "withdrawal", "transfer_in", "transfer_out"]
            },
            "description": "Tipo di transazione da filtrare (opzionale)"
          },
          {
            "in": "query",
            "name": "date_from",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data inizio yyyy-mm-dd"
          },
          {
            "in": "query",
            "name": "date_to",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data fine yyyy-mm-dd"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Numero massimo di transazioni da ritornare"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista transazioni filtrate con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "account_id": {
                            "type": "integer"
                          },
                          "transaction_type": {
                            "type": "string"
                          },
                          "amount": {
                            "type": "number",
                            "format": "float"
                          },
                          "balance_after": {
                            "type": "number",
                            "format": "float"
                          },
                          "description": {
                            "type": "string"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "account_number": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    },
                    "filters": {
                      "type": "object",
                      "properties": {
                        "account_id": {
                          "type": "string",
                          "nullable": true
                        },
                        "transaction_type": {
                          "type": "string",
                          "nullable": true
                        },
                        "date_from": {
                          "type": "string",
                          "nullable": true
                        },
                        "date_to": {
                          "type": "string",
                          "nullable": true
                        },
                        "limit": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token mancante o non valido"
          },
          "500": {
            "description": "Errore nel recupero delle transazioni"
          }
        }
      }
    },
...
    "/transactions/transfer": {
      "post": {
        "tags": ["Transactions"],
        "summary": "Esegue un bonifico tra due conti",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["from_account_id", "to_account_id", "amount"],
                "properties": {
                  "from_account_id": {
                    "oneOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "string"
                      }
                    ],
                    "example": 123,
                    "description": "Id conto mittente"
                  },
                  "to_account_id": {
                    "oneOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "string"
                      }
                    ],
                    "example": "IT60X0541234567890",
                    "description": "Id o IBAN conto destinatario"
                  },
                  "amount": {
                    "type": "number",
                    "format": "float",
                    "example": 75.0,
                    "description": "Importo del bonifico"
                  },
                  "description": {
                    "type": "string",
                    "example": "Bonifico stipendio",
                    "description": "Descrizione opzionale del bonifico"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bonifico completato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "from_account": {
                          "type": "integer"
                        },
                        "to_account": {
                          "type": "integer"
                        },
                        "amount": {
                          "type": "number",
                          "format": "float"
                        },
                        "new_from_balance": {
                          "type": "number",
                          "format": "float"
                        },
                        "new_to_balance": {
                          "type": "number",
                          "format": "float"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Errore di validazione input o bonifico non valido"
          },
          "401": {
            "description": "Token mancante o non valido"
          },
          "500": {
            "description": "Errore nel processamento del bonifico"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}